Move ComputeInputSanction into separate class
